Removing error for ignoring fields.
Added limiter to deleting in use documents.

Adding boolean tagged value for the path. 'Publish' on/off style.

+++ kintaro/kintaro.py
@@ -44,6 +44,11 @@ class Error(Exception):
     pass
 
 
+class RemoveValueError(Error):
+    """Field needs to be removed from the document."""
+    pass
+
+
 class UnknownReferenceError(Error):
     """Error finding the referenced kintaro document."""
     pass
@@ -198,6 +203,7 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
         self._env_regex_match = None
         self._env_regex_replace = r'@env.\1'
         self._removed = set()
+        self._in_use = set()
         self._id_map = {}
         self._kintaro_locale_to_locale_strings = None
         self._locale_strings_to_kintaro_locale = None
@@ -288,12 +294,6 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
         pod_paths_to_delete = set(existing_pod_paths) - set(new_pod_paths)
         self._removed = self._removed | pod_paths_to_delete
 
-    def _fix_path_none(self, key, value):
-        if self._env_regex_match and self._env_regex_match.search(key):
-            if key.startswith(('path', '$path')) and value is None:
-                return ''
-        return value
-
     def _regroup_schema(self, schema):
         names_to_fields = {}
         for field in schema:
@@ -319,10 +319,22 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
             key = '${}'.format(key)
 
         value = self._parse_field_deep(key, value, field_data, locale=locale)
-        value = self._fix_path_none(key, value)
         return key, value
 
     def _parse_field_deep(self, key, value, field_data, locale=None):
+        # For tagged environment paths it cannot be null or it falls back.
+        if self._env_regex_match and self._env_regex_match.search(key):
+            if key.startswith(('path', '$path')):
+                # If we are modifying the path with a boolean, and it is true,
+                # We want it to be 'published', so we remove the tagged value.
+                if field_data['type'] == 'BooleanField':
+                    if value == True:
+                        raise RemoveValueError()
+
+                # For a tagged environment, if the value is null it falls back.
+                if value is None:
+                    return ''
+
         # Early exit for nothings
         if value is None:
             return None
@@ -345,11 +357,17 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
                             binding.collection, filename)
                         value[idx] = self.pod.get_doc(
                             content_path, locale=locale)
+
                         # Create the doc if it does not exist.
                         # Prevent dependency problems when new docs
                         # that do not exist yet because of import order.
                         if not value[idx].exists:
                             value[idx].write(fields={})
+
+                        # Track which documents are being referenced so they
+                        # do not get deleted by accident.
+                        self._in_use.add(value[idx].pod_path)
+
                         break
         elif 'schema_fields' in field_data:
             names_to_schema_fields = self._regroup_schema(
@@ -360,7 +378,6 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
                     locale=locale)
         if single_field:
             value = value[0]
-        value = self._fix_path_none(key, value)
 
         return value
 
@@ -446,9 +463,21 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
                 continue
             raw_name = _get_base_field(name)
             field_data = names_to_schema_fields[raw_name]
-            key, value = self._parse_field(
-                name, value, field_data, locale=locale)
-            clean_fields[key] = value
+            try:
+                key, value = self._parse_field(
+                    name, value, field_data, locale=locale)
+                clean_fields[key] = value
+
+                if key.startswith('$path') and value in ('', None):
+                    # When removing a path in a tagged environment, want it to
+                    # persist to the localized value as well.
+                    if self._env_regex_match and self._env_regex_match.search(key):
+                        localization = clean_fields.get('$localization', {})
+                        localization[key.lstrip('$')] = value
+                        clean_fields['$localization'] = localization
+
+            except RemoveValueError:
+                pass
         # Populate $meta.
         if schema:
             # Strip modified info from schema.
@@ -646,6 +675,9 @@ class KintaroPreprocessor(_GoogleServicePreprocessor):
 
         # Handle deleted.
         for pod_path in self._removed:
+            if pod_path in self._in_use:
+                self.pod.logger.info('Skipping delete for in use reference -> {}'.format(pod_path))
+                continue
             self.pod.delete_file(pod_path)
             self.pod.logger.info('Deleted -> {}'.format(pod_path))
